refactor: move RecordInternal into Shard crate
