+++ lib/collection/src/collection/point_ops.rs
@@ -8,6 +8,7 @@ use futures::{StreamExt as _, TryFutureExt, TryStreamExt as _, future};
 use itertools::Itertools;
 use segment::data_types::order_by::{Direction, OrderBy};
 use segment::types::{ShardKey, WithPayload, WithPayloadInterface};
+use shard::retrieve::record_internal::RecordInternal;
 
 use super::Collection;
 use crate::operations::consistency_params::ReadConsistency;
