+++ lib/collection/src/shards/dummy_shard.rs
@@ -13,13 +13,14 @@ use segment::types::{
     ExtendedPointId, Filter, ScoredPoint, SizeStats, SnapshotFormat, WithPayload,
     WithPayloadInterface, WithVector,
 };
+use shard::retrieve::record_internal::RecordInternal;
 use tokio::runtime::Handle;
 
 use crate::operations::OperationWithClockTag;
 use crate::operations::types::{
     CollectionError, CollectionInfo, CollectionResult, CoreSearchRequestBatch,
-    CountRequestInternal, CountResult, OptimizersStatus, PointRequestInternal, RecordInternal,
-    ShardStatus, UpdateResult,
+    CountRequestInternal, CountResult, OptimizersStatus, PointRequestInternal, ShardStatus,
+    UpdateResult,
 };
 use crate::operations::universal_query::shard_query::{ShardQueryRequest, ShardQueryResponse};
 use crate::shards::shard_trait::ShardOperation;
