+++ lib/edge/src/lib.rs
@@ -1,3 +1,5 @@
+mod retrieve;
+
 use std::num::NonZero;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
