+++ lib/shard/src/lib.rs
@@ -11,5 +11,6 @@ pub mod wal;
 #[cfg(feature = "testing")]
 pub mod fixtures;
 pub mod query;
+pub mod retrieve;
 
 pub type PeerId = u64;
