+++ lib/shard/src/retrieve/mod.rs
@@ -1,0 +1,1 @@
+pub mod record_internal;
