+++ src/common/query.rs
@@ -10,6 +10,7 @@ use collection::operations::types::*;
 use collection::operations::universal_query::collection_query::*;
 use common::counter::hardware_accumulator::HwMeasurementAcc;
 use segment::types::ScoredPoint;
+use shard::retrieve::record_internal::RecordInternal;
 use storage::content_manager::errors::StorageError;
 use storage::content_manager::toc::TableOfContent;
 use storage::rbac::Access;
